requirements.txt
boto3
psycopg2-binary
python-amazon-sp-api